fix : 컴포넌트명 변경 (alarm -> notification)
