+++ src/components/index.js
@@ -26,7 +26,7 @@ import {
   Tile,
   Survey,
   MenuTab,
-  AlarmTile,
+  NotificationTile,
   ChatUserProfile,
   ProfileBox,
   VoteResultBar,
@@ -71,7 +71,7 @@ export {
   ProfileBox,
   VoteResultBar,
   MenuTab,
-  AlarmTile,
+  NotificationTile,
   CommentBox,
   ReplyCommentBox,
   BoardList,
